force cpu passthrough
reason: daily-xivo-pxe (virt-install'ed every day) sometimes triggers an
internal KVM error during Wazo installer. The error may be due to qemu
trying to emulate an unknown instruction. This attempts to tell qemu to
not emulate anything and let the instructions be interpreted by the CPU.

This started to happen since Wazo is using Debian Stretch. The current
hypothesis is that the Debian Stretch installer is using different CPU
instructions than the Jessie version.